cast to long

+++ src/main/java/fuck/manthe/nmsl/controller/AuthController.java
@@ -67,7 +67,7 @@ public class AuthController {
         try (Response response = httpClient.newCall(new Request.Builder().post(okhttp3.RequestBody.create("email=" + sharedUsername + "&password=" + this.sharedPassword + "&hwid=FUMANTHE&v=v3&t=true", MediaType.parse("application/x-www-form-urlencoded"))).url("https://www.vape.gg/auth.php").header("User-Agent", "Agent_114514").build()).execute()) {
             if (response.body() != null) {
                 if (response.isSuccessful()) {
-                    redisTemplate.opsForValue().set(Const.COLD_DOWN, System.currentTimeMillis() + randomColdDown() * 60 * 1000);
+                    redisTemplate.opsForValue().set(Const.COLD_DOWN, System.currentTimeMillis() + (long) randomColdDown() * 60 * 1000);
                 }
                 return response.body().string();
             }
